pm: rtl87x2g: restore ticks and cycles in pend stage
Restore cur_ticks and cycle_count in pend stage rather than
waiting util work_timeout_process. This is because if you wait until
work_timeout_process to restore them, there might be a situation where
after restoring nvic in work_device_resume, an interrupt is triggered
that enters the isr. At this point, ticks and cycle have not been
restored, and if the timing API is called, it will return the values
from when entering DLPS.

Signed-off-by: Zhiyuan Tang <zhiyuan_tang@realsil.com.cn>